requirements.txt
flask
playwright
pillow
tk
